requirements.txt
streamlit
openai
langchain
dotenv
tiktoken
pinecone-client
